# backend/ApiContainer/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Device Remote Management - Backend API",
    "version": "1.0.1",
    "description": "RESTful API for the Device Remote Management platform. Serves operations across SNMP, WebPA, TR69/ACS, and TR369/USP with multi-tenant isolation.",
    "contact": {
      "name": "API Support",
      "email": "api-support@devicemanagement.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "http://localhost:8080/api/v1", "description": "Local development" }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "User login (stub)",
        "description": "Validates credentials and returns JWT tokens (demo implementation).",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" } }, "required": ["email","password"] }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": {
                  "accessToken": { "type": "string" },
                  "refreshToken": { "type": "string" },
                  "user": { "type": "object" }
                } } }
            }
          },
          "400": { "description": "Invalid request" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "description": "Exchanges a valid refresh token for a new access token.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object", "properties": { "refreshToken": { "type": "string" } }, "required": ["refreshToken"] } }
          }
        },
        "responses": {
          "200": { "description": "New access token", "content": { "application/json": { "schema": { "type": "object", "properties": { "accessToken": { "type": "string" } } } } } },
          "400": { "description": "Invalid request" },
          "401": { "description": "Invalid or expired refresh token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout (revoke refresh)",
        "description": "Revokes a provided refresh token.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "refreshToken": { "type": "string" } } } } }
        },
        "responses": {
          "204": { "description": "Logged out" }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get current user profile",
        "tags": ["Users"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "summary": "Update current user profile",
        "tags": ["Users"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Updated user profile" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/devices": {
      "get": {
        "summary": "List devices (paginated)",
        "tags": ["Devices"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } }
        ],
        "responses": {
          "200": { "description": "Devices list with pagination" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create device",
        "tags": ["Devices"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["name","identifier"], "properties": {
                "name": { "type": "string" },
                "identifier": { "type": "string" },
                "mgmt_address": { "type": "string" },
                "tags": { "type": "array", "items": { "type": "string" } }
              } }
            }
          }
        },
        "responses": {
          "201": { "description": "Device created" },
          "400": { "description": "Invalid request" },
          "401": { "description": "Unauthorized" },
          "409": { "description": "Identifier conflict" }
        }
      }
    },
    "/devices/{id}": {
      "get": {
        "summary": "Get device by id",
        "tags": ["Devices"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Device" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update device",
        "tags": ["Devices"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": {
                "name": { "type": "string" },
                "identifier": { "type": "string" },
                "mgmt_address": { "type": "string" },
                "tags": { "type": "array", "items": { "type": "string" } }
              } }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated device" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" },
          "409": { "description": "Identifier conflict" }
        }
      },
      "delete": {
        "summary": "Delete device",
        "tags": ["Devices"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "List jobs (placeholder)",
        "tags": ["Jobs"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Jobs list" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "summary": "Get job by id (placeholder)",
        "tags": ["Jobs"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Job detail" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    },
...
  }
}